requirements.txt
paho-mqtt
lancedb
pandas
pyarrow
sentence-transformers
pytest
pytest-mock
numpy
